[v2] Refactor cleanup
